+++ src/Checkbox/Icon.js
@@ -1,7 +1,7 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import propTypes from 'prop-types'
 import cx from 'classnames'
-import css, { resolve } from 'styled-jsx/css'
+import { resolve } from 'styled-jsx/css'
 
 import { colors, theme } from '../theme.js'
 import { Indeterminate, Checked, Unchecked } from '../icons/Checkbox'
@@ -71,6 +71,7 @@ export const Icon = ({
             <style jsx>{`
                 div {
                     position: relative;
+                    margin: 0 6px 0 0;
                 }
 
                 .focus:before {